docs(demos and intro) updated intro, moved demos
Issue: n/a
