+++ example/src/components/pages/Introduction.css
@@ -1,0 +1,4 @@
+.intro-symbol path {
+    fill: none;
+    stroke: red
+}
